fix: resolve ESLint and Prettier issues
