+++ src/app/index.ts
@@ -1,5 +1,5 @@
 import '@/app/styles/global.css';
 
 export function startApp(rootElement: Element) {
-  console.log('app started, root:', rootElement);
+  rootElement.classList.add('app-root');
 }
