--- src/app/shared/ui/baseClasses.ts
@@ -1,14 +1,0 @@
-export const baseClasses: Record<string, string> = {
-  div: 'p-2 sm:p-4 md:p-6 w-full sm:w-2/3 md:w-1/2 text-sm sm:text-base md:text-lg',
-  p: 'p-2 sm:p-4 md:p-6 w-full sm:w-2/3 md:w-1/2 text-sm sm:text-base md:text-lg',
-  h1: 'font-bold leading-tight mb-2 text-2xl sm:text-3xl md:text-4xl',
-  h2: 'font-bold leading-tight mb-2 text-xl sm:text-2xl md:text-3xl',
-  h3: 'font-bold leading-tight mb-2 text-lg sm:text-xl md:text-2xl',
-  h4: 'font-bold leading-tight mb-2 text-base sm:text-lg md:text-xl',
-  input:
-    'w-full sm:w-1/2 md:w-1/3 px-3 py-2 border rounded-md shadow-sm text-sm focus:outline-none focus:ring-2 focus:ring-blue-400 hover:shadow-md bg-gray-50 sm:text-base sm:px-4 sm:py-2 md:text-lg md:px-5 md:py-3',
-  button:
-    'bg-sky-400 hover:bg-sky-700 p-2 w-18 sm:w-28 md:w-34 text-base sm:text-lg md:text-xl text-white rounded-2xl transition-colors duration-300 flex items-center justify-center',
-  a: 'bg-sky-400 hover:bg-sky-700 p-2 w-18 sm:w-28 md:w-34 text-base sm:text-lg md:text-xl text-white rounded-2xl transition-colors duration-300 flex items-center justify-center no-underline',
-  img: 'max-w-full h-auto object-cover',
-};
