+++ src/app/shared/ui/image/Image.ts
@@ -1,18 +1,18 @@
-import { baseClasses } from '../baseClasses';
+import { baseClasses } from '../base-classes';
 
 export function Image({
-  src = '',
+  source,
   alt = '',
   extraClasses = '',
 }: {
-  src: string;
+  source: string;
   alt?: string;
   extraClasses?: string;
 }): HTMLImageElement {
   const base = baseClasses['img'];
   const image = document.createElement('img');
   image.className = `${base} ${extraClasses}`.trim();
-  image.src = src;
+  image.src = source;
   image.alt = alt;
 
   return image;
