+++ src/app/shared/ui/input/Input.ts
@@ -1,4 +1,4 @@
-import { baseClasses } from '../baseClasses';
+import { baseClasses } from '../base-classes';
 
 export function Input({
   type = 'text',
